Improve code composition
