+++ src/products/views/ProductUpdate/consts.ts
@@ -1,0 +1,9 @@
+import { ChannelWithVariantData } from "./types";
+
+export const CHANNELS_AVAILIABILITY_MODAL_SELECTOR = "open-channels-picker";
+
+export const initialChannelWithVariantData: ChannelWithVariantData = {
+  variantsIdsToRemove: [],
+  variantsIdsToAdd: [],
+  selectedVariantsIds: []
+};
